requirements.txt
configparser
requests
beautifulsoup4
openai
python-dotenv